Moved multicore to cluster
