+++ lib/cluster.js
